feat: directory refactoring (single figure class - single cs)
